リネーム対応とStateless WIdget化
