--- lib/view/page/my_home_page.dart
@@ -1,57 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:yumemi_flutter_engineer_codecheck/l10n/app_localizations.dart';
-import 'package:yumemi_flutter_engineer_codecheck/static/wording_data.dart';
-import 'package:yumemi_flutter_engineer_codecheck/view/widget/custom_drawer.dart';
-import 'package:yumemi_flutter_engineer_codecheck/view/widget/theme_mode_select_button.dart';
-
-class MyHomePage extends ConsumerStatefulWidget {
-  const MyHomePage({super.key});
-
-  @override
-  ConsumerState<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends ConsumerState<MyHomePage> {
-  var _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final localizations = AppLocalizations.of(context);
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(localizations?.homePageTitle ?? 'Home Page'),
-        actions: const [
-          ThemeModeSelectButton(),
-        ],
-      ),
-      drawer: const CustomDrawer(),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              localizations?.counterText ?? WordingData.counterText,
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headlineMedium,
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip:
-            localizations?.incrementTooltip ?? WordingData.incrementTooltip,
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-}
